(refactor): clean up `get_elem_name`
